pyproject.toml
[project]
name = "auto-record-my-activates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dashscope>=1.24.6",
    "langchain-community>=0.3.29",
    "langgraph>=0.6.7",
    "mss>=10.1.0",
    "pillow>=11.3.0",
    "pyqt6>=6.9.1",
    "python-dotenv>=1.1.1",
]
